Removed the time splitting construction, refactored the code to use entries
