+++ lib/physiqual/workers.rb
@@ -1,0 +1,7 @@
+require 'httparty'
+require 'physiqual/workers/cache_worker'
+
+module Physiqual
+  module Workers
+  end
+end
