+++ spec/lib/physiqual/cassandra_connection_spec.rb
@@ -40,6 +40,17 @@ module Physiqual
       end
 
       describe 'insert' do
+        it "should insert data in the database in batches" do
+          expect(true).to eq false
+        end
+
+        it "should be able to work with nil values" do
+          expect(true).to eq false
+        end
+
+        it 'should convert values to big decimal values if needed' do
+
+        end
       end
 
       describe 'query_heart_rate' do
