+++ spec/lib/physiqual/data_services/cassandra_data_service_spec.rb
@@ -1,0 +1,6 @@
+module Physiqual
+  require 'rails_helper'
+  describe CassandraDataService do
+
+  end
+end
