+++ spec/lib/physiqual/workers/cache_worker_spec.rb
@@ -1,0 +1,8 @@
+module Physiqual
+  module Workers
+    require 'rails_helper'
+    describe CacheWorker do
+
+    end
+  end
+end
